requirements.txt
click==8.1.7
et-xmlfile==1.1.0
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
pillow==10.3.0
setuptools==69.5.1
wheel==0.43.0
